ImageProcessingScripts/Image To Text Converter/requirements.txt
Pillow==8.4.0
PySimpleGUI==4.55.1
pytesseract==0.3.8
